+++ packages/jest-transform/src/index.ts
@@ -11,9 +11,12 @@ export {
 } from './ScriptTransformer';
 export {default as shouldInstrument} from './shouldInstrument';
 export type {
+  CacheKeyOptions,
   Transformer,
   ShouldInstrumentOptions,
   Options as TransformationOptions,
+  TransformOptions,
   TransformResult,
+  TransformedSource,
 } from './types';
 export {default as handlePotentialSyntaxError} from './enhanceUnexpectedTokenMessage';