+++ scripts/checkCopyrightHeaders.js
@@ -102,6 +102,8 @@ const CUSTOM_IGNORED_PATTERNS = [
   '^packages/expect/src/jasmineUtils\\.ts$',
   '^packages/jest-config/src/vendor/jsonlint\\.js$',
   '^packages/jest-diff/src/cleanupSemantic\\.ts$',
+  '^website/static/css/code-block-buttons\\.css$',
+  '^website/static/js/code-block-buttons\\.js',
 ].map(createRegExp);
 
 const IGNORED_PATTERNS = [